PSM libs: p4. Fix signed vs unsigned comparison warning
shell-encryption and private_membership leaked some configs
which disable some default warnings.

In order to re-enable all these warnings for chrome/browser/*
in this CL crrev.com/c/3303510, all warnings need to be fixed.

This CL is fixing the comparison between two variables to have
the same type (i.e. either both are signed or unsigned integers).

Bug: chromium:1212837, b:201763017
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3470395
Commit-Queue: Amr Aboelkher <amraboelkher@chromium.org>
Cr-Commit-Position: refs/heads/main@{#972436}

+++ chrome/browser/ash/policy/remote_commands/device_command_screenshot_job.h
@@ -26,7 +26,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace policy {
 
 // The first element represents screen index, the second presents the PNG data.
-using ScreenshotData = std::pair<int, scoped_refptr<base::RefCountedMemory>>;
+using ScreenshotData = std::pair<size_t, scoped_refptr<base::RefCountedMemory>>;
 
 using OnScreenshotTakenCallback =
     base::OnceCallback<void(scoped_refptr<base::RefCountedMemory>)>;
